@@ -1,25 +1,24 @@
 source 'https://rubygems.org'
 ruby '2.0.0'
 
-gem 'rails', '3.2.13'
+gem 'rails', '4.0.0.rc2'
 gem 'pg'
 
 gem 'airbrake'
 gem 'bcrypt-ruby'
+gem 'bootstrap-sass-rails'
 gem 'capistrano'
+gem 'coffee-rails', '~> 4.0.0'
+gem 'compass-rails', '>= 2.0.alpha.0'
 gem 'haml-rails'
 gem 'jquery-rails'
 gem 'onload_js-rails'
-gem 'simple_form'
-gem 'strong_parameters'
-
-group :assets do
-  gem 'bootstrap-sass-rails'
-  gem 'coffee-rails'
-  gem 'compass-rails'
-  gem 'sass-rails'
-  gem 'uglifier'
-end
+gem 'pry'
+gem 'sass-rails', '~> 4.0.0.rc2'
+gem 'simple_form', '>= 3.0.0.rc'
+gem 'uglifier', '>= 1.3.0'
+gem 'turbolinks'
+gem 'jbuilder', '~> 1.2'
 
 group :development, :test do
   gem 'awesome_print', require: 'ap'
